@@ -107,7 +107,6 @@ if __name__ == "__main__":
         processed_basepath = Path(FLAGS.processed_basedir)
         processed_basepath.mkdir(exist_ok=True, parents=True)
 
-        loop = asyncio.get_event_loop()
-        loop.run_until_complete(process_hymns())
+        asyncio.run(process_hymns())
 
     app.run(main)
